vibefactory/requirements.txt
# Core dependencies
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# AI/ML
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain>=0.1.0

# Data processing
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
loguru>=0.7.0

# Development
types-pyyaml>=6.0.12
types-python-dateutil>=2.8.19
types-requests>=2.31.0.10
